+++ fakedb/include/oltp.h
@@ -2,8 +2,9 @@
 #define _OLTP_H_
 
 #include <ctime>
+#include <stdint.h>                    //for uint64_t
 
-extern void newOrderRandom(time_t now, int w_id);
-extern void deliveryRandom(time_t now, int w_id);
+extern void newOrderRandom(time_t now, uint64_t w_id);
+extern void deliveryRandom(time_t now, uint64_t w_id);
 
 #endif //_OLTP_H_
